File system object & downcasting (#173)
* removed read_only flag and integrated into config object
* Removed size_t and ssize_t from src/
* explicit downcast 32 > 64 bit integer & check

+++ src/common/serializer.cpp
@@ -3,20 +3,20 @@
 using namespace duckdb;
 using namespace std;
 
-Serializer::Serializer(size_t maximum_size)
+Serializer::Serializer(uint64_t maximum_size)
     : Serializer(unique_ptr<uint8_t[]>(new uint8_t[maximum_size]), maximum_size) {
 }
 
-Serializer::Serializer(unique_ptr<uint8_t[]> data, size_t size) : maximum_size(size), data(data.get()) {
+Serializer::Serializer(unique_ptr<uint8_t[]> data, uint64_t size) : maximum_size(size), data(data.get()) {
 	blob.size = 0;
 	blob.data = move(data);
 }
 
-Serializer::Serializer(uint8_t *data) : maximum_size((size_t)-1), data(data) {
+Serializer::Serializer(uint8_t *data) : maximum_size((uint64_t)-1), data(data) {
 	blob.size = 0;
 }
 
-Deserializer::Deserializer(uint8_t *ptr, size_t data) : ptr(ptr), endptr(ptr + data) {
+Deserializer::Deserializer(uint8_t *ptr, uint64_t data) : ptr(ptr), endptr(ptr + data) {
 }
 
 template <> string Deserializer::Read() {
