+++ src/include/common/vector_operations/unary_loops.hpp
@@ -21,10 +21,11 @@ inline void UNARY_TYPE_CHECK(Vector &input, Vector &result) {
 }
 
 template <class LEFT_TYPE, class RESULT_TYPE, class OP>
-static inline void unary_loop_function(LEFT_TYPE *__restrict ldata, RESULT_TYPE *__restrict result_data, size_t count,
+static inline void unary_loop_function(LEFT_TYPE *__restrict ldata, RESULT_TYPE *__restrict result_data, uint64_t count,
                                        sel_t *__restrict sel_vector) {
 	ASSERT_RESTRICT(ldata, ldata + count, result_data, result_data + count);
-	VectorOperations::Exec(sel_vector, count, [&](size_t i, size_t k) { result_data[i] = OP::Operation(ldata[i]); });
+	VectorOperations::Exec(sel_vector, count,
+	                       [&](uint64_t i, uint64_t k) { result_data[i] = OP::Operation(ldata[i]); });
 }
 
 template <class LEFT_TYPE, class RESULT_TYPE, class OP> void templated_unary_loop(Vector &input, Vector &result) {
@@ -39,14 +40,15 @@ template <class LEFT_TYPE, class RESULT_TYPE, class OP> void templated_unary_loo
 
 template <class LEFT_TYPE, class RESULT_TYPE, class OP>
 static inline void unary_loop_process_null_function(LEFT_TYPE *__restrict ldata, RESULT_TYPE *__restrict result_data,
-                                                    size_t count, sel_t *__restrict sel_vector, nullmask_t &nullmask) {
+                                                    uint64_t count, sel_t *__restrict sel_vector,
+                                                    nullmask_t &nullmask) {
 	ASSERT_RESTRICT(ldata, ldata + count, result_data, result_data + count);
 	if (nullmask.any()) {
 		VectorOperations::Exec(sel_vector, count,
-		                       [&](size_t i, size_t k) { result_data[i] = OP::Operation(ldata[i], nullmask[i]); });
+		                       [&](uint64_t i, uint64_t k) { result_data[i] = OP::Operation(ldata[i], nullmask[i]); });
 	} else {
 		VectorOperations::Exec(sel_vector, count,
-		                       [&](size_t i, size_t k) { result_data[i] = OP::Operation(ldata[i], false); });
+		                       [&](uint64_t i, uint64_t k) { result_data[i] = OP::Operation(ldata[i], false); });
 	}
 }
