+++ src/include/execution/operator/aggregate/physical_window.hpp
@@ -34,7 +34,7 @@ public:
 	PhysicalWindowOperatorState(PhysicalOperator *child) : PhysicalOperatorState(child), position(0) {
 	}
 
-	size_t position;
+	uint64_t position;
 	ChunkCollection tuples;
 	ChunkCollection window_results;
 };
