+++ src/include/planner/operator/logical_comparison_join.hpp
@@ -25,13 +25,13 @@ public:
 
 	static unique_ptr<LogicalOperator> CreateJoin(JoinType type, unique_ptr<LogicalOperator> left_child,
 	                                              unique_ptr<LogicalOperator> right_child,
-	                                              unordered_set<size_t> &left_bindings,
-	                                              unordered_set<size_t> &right_bindings,
+	                                              unordered_set<uint64_t> &left_bindings,
+	                                              unordered_set<uint64_t> &right_bindings,
 	                                              vector<unique_ptr<Expression>> &expressions);
-	size_t ExpressionCount() override;
-	Expression *GetExpression(size_t index) override;
+	uint64_t ExpressionCount() override;
+	Expression *GetExpression(uint64_t index) override;
 	void ReplaceExpression(std::function<unique_ptr<Expression>(unique_ptr<Expression> expression)> callback,
-	                       size_t index) override;
+	                       uint64_t index) override;
 
 	string ParamsToString() const override;
 };
