+++ src/include/planner/table_binding.hpp
@@ -28,7 +28,7 @@ enum class BindingType : uint8_t { TABLE = 0, SUBQUERY = 1, TABLE_FUNCTION = 2 }
 //! A Binding represents a binding to a table, table-producing function or subquery with a specified table index. Used
 //! in the binder.
 struct Binding {
-	Binding(BindingType type, const string &alias, size_t index) : type(type), alias(alias), index(index) {
+	Binding(BindingType type, const string &alias, uint64_t index) : type(type), alias(alias), index(index) {
 	}
 	virtual ~Binding() {
 	}
@@ -40,7 +40,7 @@ struct Binding {
 
 	BindingType type;
 	string alias;
-	size_t index;
+	uint64_t index;
 };
 
 //! Represents a binding to a base table
@@ -56,7 +56,7 @@ struct TableBinding : public Binding {
 
 //! Represents a binding to a subquery
 struct SubqueryBinding : public Binding {
-	SubqueryBinding(const string &alias, SubqueryRef &ref, BoundQueryNode &subquery, size_t index);
+	SubqueryBinding(const string &alias, SubqueryRef &ref, BoundQueryNode &subquery, uint64_t index);
 
 	bool HasMatchingBinding(const string &column_name) override;
 	BindResult Bind(ColumnRefExpression &colref, uint32_t depth) override;
@@ -66,12 +66,12 @@ struct SubqueryBinding : public Binding {
 	//! Column names of the subquery
 	vector<string> names;
 	//! Name -> index for the names
-	unordered_map<string, size_t> name_map;
+	unordered_map<string, uint64_t> name_map;
 };
 
 //! Represents a binding to a table-producing function
 struct TableFunctionBinding : public Binding {
-	TableFunctionBinding(const string &alias, TableFunctionCatalogEntry *function, size_t index);
+	TableFunctionBinding(const string &alias, TableFunctionCatalogEntry *function, uint64_t index);
 
 	bool HasMatchingBinding(const string &column_name) override;
 	BindResult Bind(ColumnRefExpression &colref, uint32_t depth) override;
